config: fix compile error on non-linux platform
